Refactor: interface 파일 정리
