+++ src/components/modals/DeleteCardModal/DeleteCardModal.tsx
@@ -1,7 +1,7 @@
-import { CardProps } from "@/types";
+import { CardInterface } from "@/types";
 import styles from "./DeleteCardModal.module.scss";
 
-function DeleteCardModal({ card }: CardProps) {
+function DeleteCardModal({ card }: { card: CardInterface }) {
   return (
     <div className={styles["modal-content"]}>
       <h2 className={styles["modal-title"]}>링크 삭제</h2>
