--- src/types/CardProps.tsx
@@ -1,5 +1,0 @@
-import CardInterface from "./CardInterface";
-
-export default interface CardProps {
-  card: CardInterface;
-}
