--- src/types/FolderProps.tsx
@@ -1,5 +1,0 @@
-import FolderInterface from "./FolderInterface";
-
-export default interface FolderProps {
-  folder: FolderInterface;
-}
