--- src/types/NavProps.tsx
@@ -1,6 +1,0 @@
-import { UserInterface } from ".";
-
-export default interface NavProps {
-  profile?: UserInterface;
-  isSticky?: boolean;
-}
