--- src/types/SharedProps.tsx
@@ -1,7 +1,0 @@
-import SharedFolderInterface from "./SharedFolderInterface";
-import SharedUserInterface from "./SharedUserInterface";
-
-export default interface SharedProps {
-  sharedFolder?: SharedFolderInterface;
-  sharedUser?: SharedUserInterface;
-}
